+++ tests/units/feature_extraction/test_feature_calculations.py
@@ -322,6 +322,38 @@ class FeatureCalculationTestCase(TestCase):
         self.assertEqualOnAllArrayTypes(cid_ce, [0.5, 3.5, 7.5], 5, normalize=False)
         self.assertEqualOnAllArrayTypes(cid_ce, [-4.33, -1.33, 2.67], 5, normalize=False)
 
+    def test_lempel_ziv_complexity(self):
+        self.assertAlmostEqualOnAllArrayTypes(lempel_ziv_complexity, [1, 1, 1], 2./3, bins=2)
+        self.assertAlmostEqualOnAllArrayTypes(lempel_ziv_complexity, [1, 1, 1], 2./3, bins=5)
+
+        self.assertAlmostEqualOnAllArrayTypes(lempel_ziv_complexity, [1, 1, 1, 1, 1, 1, 1],
+                                              0.4285714285, bins=2)
+        self.assertAlmostEqualOnAllArrayTypes(lempel_ziv_complexity, [1, 1, 1, 2, 1, 1, 1],
+                                              0.5714285714, bins=2)
+
+        self.assertAlmostEqualOnAllArrayTypes(lempel_ziv_complexity,
+                                              [-1, 4.3, 5, 1, -4.5, 1, 5, 7, -3.4, 6],
+                                              0.8, bins=10)
+        self.assertAlmostEqualOnAllArrayTypes(lempel_ziv_complexity,
+                                              [-1, np.nan, 5, 1, -4.5, 1, 5, 7, -3.4, 6],
+                                              0.4, bins=10)
+
+    def test_fourier_entropy(self):
+        self.assertAlmostEqualOnAllArrayTypes(fourier_entropy, [1, 2, 1], 0.693147180, bins=2)
+        self.assertAlmostEqualOnAllArrayTypes(fourier_entropy, [1, 2, 1], 0.693147180, bins=5)
+
+        self.assertAlmostEqualOnAllArrayTypes(fourier_entropy, [1, 1, 2, 1, 1, 1, 1],
+                                              0.5623351446188083, bins=5)
+        self.assertAlmostEqualOnAllArrayTypes(fourier_entropy, [1, 1, 1, 1, 2, 1, 1],
+                                              1.0397207708399179, bins=5)
+
+        self.assertAlmostEqualOnAllArrayTypes(fourier_entropy,
+                                              [-1, 4.3, 5, 1, -4.5, 1, 5, 7, -3.4, 6],
+                                              1.5607104090414063, bins=10)
+        self.assertIsNanOnAllArrayTypes(fourier_entropy,
+                                        [-1, np.nan, 5, 1, -4.5, 1, 5, 7, -3.4, 6],
+                                        bins=10)
+
     def test_ratio_beyond_r_sigma(self):
 
         x = [0, 1] * 10 + [10, 20, -30]  # std of x is 7.21, mean 3.04